pyproject.toml
[build-system]
requires = ["setuptools>=75.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "marimo-flow"
version = "0.3.0"
description = "Reactive marimo notebooks for ML with MLflow tracking, PINA physics-informed neural networks, and a multi-agent team on pydantic-graph + Ollama Cloud."
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
license-files = ["LICENSE"]
authors = [
    {name = "Björn Bethge", email = "bjoern.bethge@gmail.com"},
]
keywords = [
    "marimo",
    "mlflow",
    "pina",
    "physics-informed-neural-networks",
    "pde-solver",
    "reactive-notebooks",
    "machine-learning",
    "pydantic-ai",
    "pydantic-graph",
    "ollama",
    "mcp",
    "model-context-protocol",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "marimo[lsp,recommended,sql,mcp]>=0.23.1",
    "mlflow[genai,mcp]>=3.11.1",
    "optuna>=4.8.0",
    "pina-mathlab>=0.2.6",
    "transformers[accelerate,sentencepiece,tiktoken]>=5.5.4",
    # security: explicit minimums for vulnerable transitive deps
    "authlib>=1.6.9",
    "pillow>=12.2.0",
    "cryptography>=46.0.7",
    "langgraph>=1.1.8",
    "werkzeug>=3.1.8",
    "sqlparse>=0.5.5",
    "diskcache>=5.6.3",
    "protobuf>=6.33.6",
    "python-multipart>=0.0.22",
    "ujson>=5.12.0",
    "pyasn1>=0.6.3",
    "pyjwt>=2.12.1",
    "flask>=3.1.3",
    "optuna-integration[pytorch-lightning]>=4.8.0",
    "plotly[express,kaleido]>=6.7.0",
    "pydantic-ai-slim[a2a,ag-ui]>=1.84.1",
    "pyyaml>=6.0.3",
    "python-dotenv>=1.2.2",
    "typer>=0.24.1",
    "sympy>=1.14.0",
]

[project.optional-dependencies]
all = [
    "torch>=2.11.0",
    "torch-geometric>=2.7.0",
]

[dependency-groups]
dev = [
    "black>=26.3.1",
    "mypy>=1.20.1",
    "pre-commit>=4.5.1",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.15.11",
]
docs = [
    "sphinx>=9.0.4",
    "sphinx-rtd-theme>=3.1.0",
]

[project.urls]
Homepage = "https://github.com/synapticore-io/marimo-flow"
Repository = "https://github.com/synapticore-io/marimo-flow"

[project.scripts]
marimo-flow = "marimo_flow.cli:main"

[tool.setuptools]
include-package-data = true
zip-safe = false

[tool.setuptools.packages.find]
where = ["src"]
include = ["marimo_flow*"]
exclude = ["tests*", "docs*", "examples*", "raw*", "models*", "mlruns*"]

[tool.setuptools.package-data]
marimo_flow = ["py.typed", "*.yaml", "*.yml", "*.json", "*.css", "*.mp4"]

[tool.uv]
link-mode = "copy"
# Override mlflow's `fastmcp<3` pin to pick up security fixes (CVE-2026-32871,
# CVE-2026-27124, CVE-2025-64340). Revisit once mlflow supports fastmcp 3.
override-dependencies = ["fastmcp>=3.2.0"]

[tool.ruff]
line-length = 88
target-version = "py311"
extend-exclude = [".claude/Skills/*/examples"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
# marimo cells render the last bare expression as cell output
"examples/*" = ["B018"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.uv.sources]
